[013] Migrate StringValues extensions into consolidated file
- Move IsNull, Length, and ToString regions from legacy files to StringValuesExtensions.cs

- Delete legacy StringValuesIsNullExtensions.cs, StringValuesLengthExtensions.cs, StringValuesToStringExtensions.cs

Refs: ticket-013, DECISIONS-DotExtensions-ms-extensions-collapse.md#D002
