+++ DotExtensions/MsExtensions/Primitives/StringValuesExtensions.cs
@@ -25,12 +25,92 @@
 namespace DotExtensions.MsExtensions.Primitives;
 
 /// <summary>
-/// Provides extension methods for converting StringValues to string representations with various separators.
+/// Provides extension methods for working with <see cref="StringValues"/> to determine
+/// if they are null, empty, or contain only null or whitespace values.
 /// </summary>
-public static class StringValuesToStringExtensions
+public static class StringValuesExtensions
 {
     /// <summary>
-    /// Provides extension methods for converting StringValues to string representations with various separators.
+    /// Provides extension methods for null checks on <see cref="StringValues"/>.
+    /// </summary>
+    extension(StringValues strValues)
+    {
+        /// <summary>
+        /// Whether this <see cref="StringValues"/> is empty.
+        /// </summary>
+        /// <returns>True if it is empty, false otherwise.</returns>
+        public bool IsEmpty => strValues.Equals(StringValues.Empty);
+    }
+    
+    extension(StringValues)
+    {
+        /// <summary>
+        /// Determines whether a <see cref="StringValues"/> contains any strings that are null or whitespace./>
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns>True if any of the strings is WhiteSpace or null.</returns>
+        public static bool IsNullOrWhiteSpace(StringValues? other)
+        {
+            if (other is null)
+                return true;
+            
+            return StringValues.IsWhiteSpace((StringValues)other);
+        }
+
+        /// <summary>
+        /// Determines whether a <see cref="StringValues"/> contains only strings that consist entirely of whitespace characters.
+        /// </summary>
+        /// <param name="other">The <see cref="StringValues"/> to check for whitespace characters.</param>
+        /// <returns>True if all strings in the <see cref="StringValues"/> consist entirely of whitespace characters; otherwise, false.</returns>
+        private static bool IsWhiteSpace(StringValues other)
+        {
+            if (other.Count == 0)
+                return false;
+            
+            bool[] vals = new bool[other.Count];
+
+            for (int index = 0; index < other.Count; index++)
+            {
+                string? val = other[index];
+                
+                vals[index] = string.IsNullOrWhiteSpace(val);
+            }
+
+            return vals.Any(x => x);
+        }
+    }
+    
+    #region Length
+
+    /// <param name="stringValues">The <see cref="StringValues"/> object to search.</param>
+    extension(StringValues stringValues)
+    {
+        /// <summary>
+        /// The total length of all strings in a <see cref="StringValues"/> object combined.
+        /// </summary>
+        public int TotalLength
+        {
+            get
+            {
+                int length = 0;
+
+                foreach (string? value in stringValues)
+                {
+                    if(value is not null)
+                        length += value.Length;
+                }
+
+                return length;
+            }  
+        }
+    }
+    
+    #endregion
+    
+    #region ToString
+
+    /// <summary>
+    /// Provides extension methods for converting <see cref="StringValues"/> to string representations with various separators.
     /// </summary>
     extension(StringValues strValues)
     {
@@ -99,4 +179,6 @@ public static class StringValuesToStringExtensions
             return output;
         }
     }
+    
+    #endregion
 }
